backend/requirements.txt
aiohttp==3.8.3
aiohttp-middlewares==2.2.0
aiohttp-tus==1.0.0
aiosignal==1.3.1
anyio==3.6.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bidict==0.22.1
bz2file==0.98
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==2.1.1
ciseau==1.0.1
click==8.1.3
decorator==5.1.1
epub==0.5.2
epub-conversion==1.0.15
executing==1.2.0
fastapi==0.89.1
fastapi-socketio==0.0.9
filelock==3.9.0
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
huggingface-hub==0.0.8
idna==2.10
ipython==8.8.0
jedi==0.18.2
joblib==1.2.0
matplotlib-inline==0.1.6
multidict==6.0.4
nltk==3.5
numpy==1.24.1
packaging==23.0
pandas==1.2.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.36
protobuf==3.20.1
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.4
Pygments==2.14.0
PyPDF2==1.26.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-engineio==4.3.4
python-multipart==0.0.5
python-socketio==5.7.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.24.0
sacremoses==0.0.53
sentencepiece==0.1.97
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
starlette==0.22.0
tika==1.24
tokenizers==0.10.3
torch==1.8.1
tqdm==4.55.1
traitlets==5.8.1
transformers==4.6.1
typing_extensions==4.4.0
urllib3==1.25.11
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
wcwidth==0.2.6
websockets==10.4
whisper==1.1.10
xml-cleaner==2.0.4
yarl==1.8.2
